Reorganized menu index for readability. Moved tests into their own folders.
