+++ lib/core_modules/menu/index.js
@@ -6,7 +6,12 @@ var _ = require('lodash'),
 
 function Menus() {}
 
-Menus.prototype.add = function (options) {
+Menus.prototype.add = add;
+Menus.prototype.get = get;
+
+module.exports = Menus;
+
+function add(options) {
   if (arguments.length === 0) return this;
   if (options instanceof Array) {
     options.forEach(Menus.prototype.add.bind(this));
@@ -24,16 +29,16 @@ Menus.prototype.add = function (options) {
 
   options = _.assign({
       path: 'main',
-      roles: ['anonymous'],
+      roles: ['anonymous']
     },
     options);
   options.path = options.path.replace(/^\//, '');
   var item = allMenus.findOrCreate(options.path.split('/'));
   item.add(new MenuItem(options));
   return this;
-};
+}
 
-Menus.prototype.get = function (options) {
+function get(options) {
   options = options || {};
   options.menu = options.menu || 'main';
   options.roles = options.roles || ['anonymous'];
@@ -46,10 +51,8 @@ Menus.prototype.get = function (options) {
   }
   var ret = sm.get(options.roles);
   return ret ? options.defaultMenu.concat(ret.submenus.map(mapDoStrip)) : options.defaultMenu;
-};
+}
 
 function mapDoStrip(v) {
   return v ? v.strip() : undefined;
 }
-
-module.exports = Menus;
